refactor: admin_company_controller.goを責務ごとに3ファイルに分割
1054行の巨大コントローラを以下の3ファイルに分割:
- admin_company_controller.go: struct/constructor/setters, CRUD, gBiz, ヘルパー (~360行)
- admin_company_fetch_controller.go: AI取得系エンドポイント (~570行)
- admin_company_catalog_controller.go: L1カタログ/バッチ処理 (~145行)

ロジック変更なし。全テスト通過。

closes #725
